feat(experiments): show Replay vision promo on the Recordings tab
Generated-By: PostHog Code
Task-Id: f1ec758f-0345-445f-836d-d9c1d0de641a
